+++ flink-runtime/src/main/java/org/apache/flink/streaming/api/graph/StreamGraphHasher.java
@@ -31,4 +31,12 @@ public interface StreamGraphHasher {
      * didn't change.
      */
     Map<Integer, byte[]> traverseStreamGraphAndGenerateHashes(StreamGraph streamGraph);
+
+    /**
+     * Generates a hash for {@link StreamNode} with the specified stream node id in the {@link
+     * StreamGraph}. This hash is stored in the provided map and can be used to uniquely identify
+     * the {@link StreamNode} across job submissions, assuming its configuration remains unchanged.
+     */
+    boolean generateHashesByStreamNodeId(
+            int streamNodeId, StreamGraph streamGraph, Map<Integer, byte[]> hashes);
 }
